nz-skill-runner/requirements.txt
# Список Python-зависимостей с закрепленными версиями для воспроизводимости
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
python-frontmatter==1.0.0
docker==7.1.0
requests==2.31.0